+++ task-viewer/src/components/TaskCard/TaskCard.css
@@ -49,3 +49,10 @@
   justify-content: flex-end;
   align-items: center;
 }
+
+.task-btn{
+  background: #f2f2f2;
+  border:#16171d3a solid;
+  border-radius: 10px 10px 10px 10px;
+  margin-right: 65px;
+}
