+++ task-viewer/src/components/badge/Badge.jsx
@@ -1,4 +1,5 @@
 import "./Badge.css";
+import React from "react";
 
 function Badge(props) {
   return (
